Fixed errors with header files
